Wed, Sep 7, 2022, 5:15 AM +00:00

+++ readability/readability.c
@@ -44,7 +44,7 @@ int main(void)
    }
    else
    {
-    printf("Grade %i\n", round(index));
+    printf("Grade %d\n", (int)round(index));
    }
 }
